Refactor MCP input record helpers
